Remove aggregators (#690)
Primarily, this removes aggregators, which have been deprecated since
1.4 and have been replaced by combiners. This removal sets us up to
provide aggregators as a separate jar artifact, outside of the main
Accumulo build, if there is a continued demand for them. In future, all
such "user library" type add-ons to Accumulo could be provided as
separate, optional component jars.

Also cleanup other minor deprecated items.

Also create a delegating ConnectorImpl to allow removal of old deprecated
methods from new AccumuloClientImpl. This ConnectorImpl reuses the
implementation moved to AccumuloClientImpl via a delegate field, but
keeps the old ConnectorImpl methods that weren't carried over to the new
interface.

+++ core/src/main/java/org/apache/accumulo/core/client/impl/ThriftScanner.java
@@ -231,7 +231,7 @@ public class ThriftScanner {
     return (long) (Math.min(millis * 2, maxSleep) * (.9 + secureRandom.nextDouble() / 5));
   }
 
-  public static List<KeyValue> scan(ClientContext context, ScanState scanState, int timeOut)
+  public static List<KeyValue> scan(ClientContext context, ScanState scanState, long timeOut)
       throws ScanTimedOutException, AccumuloException, AccumuloSecurityException,
       TableNotFoundException {
     TabletLocation loc = null;
